Streamline cli module
