+++ tests/integration/test_explorer.py
@@ -3,7 +3,7 @@ import os.path
 import pandas as pd
 from click.testing import CliRunner
 
-from silicone.cli import plot_emission_correlations
+from silicone.cli import plot_emission_correlations_cli
 
 
 def _make_input_file(df, dpath):
@@ -35,7 +35,7 @@ def test_plot_emission_correlations(check_aggregate_df, tmpdir, caplog):
     # caplog can capture any logging calls (if we add any in future)
     with caplog.at_level("DEBUG"):
         result = runner.invoke(
-            plot_emission_correlations,
+            plot_emission_correlations_cli,
             [
                 input_file,
                 "--output-dir",
@@ -88,7 +88,7 @@ def test_plot_emission_correlations_no_quantiles_or_model_colours(
     # caplog can capture any logging calls (if we add any in future)
     with caplog.at_level("DEBUG"):
         result = runner.invoke(
-            plot_emission_correlations,
+            plot_emission_correlations_cli,
             [
                 input_file,
                 "--output-dir",
@@ -126,7 +126,7 @@ def test_plot_emission_correlations_no_output(check_aggregate_df, tmpdir, caplog
     # caplog can capture any logging calls (if we add any in future)
     with caplog.at_level("DEBUG"):
         result = runner.invoke(
-            plot_emission_correlations,
+            plot_emission_correlations_cli,
             [
                 input_file,
                 "--years",
@@ -164,7 +164,7 @@ def test_plot_emission_correlations_no_co2_emms(check_aggregate_df, tmpdir, capl
     # caplog can capture any logging calls (if we add any in future)
     with caplog.at_level("DEBUG"):
         result = runner.invoke(
-            plot_emission_correlations,
+            plot_emission_correlations_cli,
             [
                 input_file,
                 "--output-dir",
